Move UrbanMscParams to top-level em directory
