+++ src/celeritas/em/UrbanMscParams.cc
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/em/msc/UrbanMscParams.cc
+//! \file celeritas/em/UrbanMscParams.cc
 //---------------------------------------------------------------------------//
 #include "UrbanMscParams.hh"
