+++ src/celeritas/global/alongstep/AlongStepGeneralLinearAction.cc
@@ -16,7 +16,7 @@
 #include "corecel/sys/MultiExceptionHandler.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/em/FluctuationParams.hh"
-#include "celeritas/em/msc/UrbanMscParams.hh"
+#include "celeritas/em/UrbanMscParams.hh"
 #include "celeritas/global/CoreTrackData.hh"
 #include "celeritas/global/KernelContextException.hh"
 #include "celeritas/phys/PhysicsParams.hh"
