+++ test/celeritas/GeantTestBase.cc
@@ -11,7 +11,7 @@
 
 #include "celeritas_cmake_strings.h"
 #include "corecel/io/Logger.hh"
-#include "celeritas/em/msc/UrbanMscParams.hh"
+#include "celeritas/em/UrbanMscParams.hh"
 #include "celeritas/ext/GeantImporter.hh"
 #include "celeritas/ext/GeantPhysicsOptions.hh"
 #include "celeritas/ext/GeantSetup.hh"
